new update

+++ app/Http/Controllers/ProfileController.php
@@ -31,7 +31,7 @@ class ProfileController extends Controller
         if (Auth::user()->role_id == 1) {
             return view('profile.create');
         }
-        return redirect()->route('profile.index')->with('info', "Ooops!Vous n'avez pas l'autorisation de modifier ce compte.");
+        return redirect()->route('profile.index')->with('info', "Ooops!Vous n'avez pas l'autorisation d'ajouter un profile.");
     }
 
     /**
@@ -63,10 +63,10 @@ class ProfileController extends Controller
     public function edit(Profile $profile)
     {
         // dump($profile);
-        if (Auth::user()->role_id == 1 || Auth::user()->role_id == 2 && $profile->role_id !== 1 ) {
+        if (Auth::user()->role_id == 1 || Auth::user()->role_id == 2 && $profile->role_id !== 1) {
             return view('profile.edit', compact('profile'));
         };
-        return redirect()->route('profile.index')->with('info', "Ooops!Vous n'avez pas l'autorisation de modifier ce compte.");
+        return redirect()->route('profile.index')->with('info', "Ooops!Vous n'avez pas l'autorisation de modifier les comptes 'Admin'");
     }
 
     /**
@@ -74,7 +74,9 @@ class ProfileController extends Controller
      */
     public function update(ProfileRequest $request, Profile $profile)
     {
-        if (Auth::user()->role_id == 1 || Auth::user()->role_id == 2 && $profile->role_id !== 1 ) {
+        if (Auth::user()->role_id == 1 || Auth::user()->role_id == 2 && $profile->role_id !== 1) {
+            // $form['password'] = Hash::make($request->password);
+            $request['password']=Hash::make($request->password);
             $profile->fill($request->post())->save();
             return redirect()->route('profile.index')->with('info', 'Le compte a été bien modifié');
         }
@@ -86,10 +88,10 @@ class ProfileController extends Controller
      */
     public function destroy(Profile $profile)
     {
-        if (Auth::user()->role_id == 1) {
+        if (Auth::user()->role_id == 1 && $profile->id !== 1) {
             $profile->delete();
             return redirect()->route('profile.index')->with('danger', 'Le compte a été bien supprimé');
         };
-        abort(403);
+        return redirect()->route('profile.index')->with('danger', 'Interdit de supprimer ce profile');
     }
 }
